code refactored - sonar issues
